+++ golem-worker-service/src/service/worker.rs
@@ -174,7 +174,7 @@ pub trait WorkerService {
     async fn complete_promise(
         &self,
         worker_id: &WorkerId,
-        oplog_id: i32,
+        oplog_id: u64,
         data: Vec<u8>,
     ) -> Result<bool, WorkerServiceBaseError>;
 
@@ -766,7 +766,7 @@ impl WorkerService for WorkerServiceDefault {
     async fn complete_promise(
         &self,
         worker_id: &WorkerId,
-        oplog_id: i32,
+        oplog_id: u64,
         data: Vec<u8>,
     ) -> Result<bool, WorkerServiceBaseError> {
         let promise_id = PromiseId {
@@ -1024,7 +1024,7 @@ impl WorkerService for WorkerServiceNoOp {
     async fn complete_promise(
         &self,
         _worker_id: &WorkerId,
-        _oplog_id: i32,
+        _oplog_id: u64,
         _data: Vec<u8>,
     ) -> Result<bool, WorkerServiceBaseError> {
         Ok(true)
